layout: move dialogs to features/modals/ to prep for Redux
The recommended directory layout [1] places Redux-connected components
in the `src/features/` tree. Move Dialog components into the
Redux-recommended directory before adding Redux integration.

[1] https://redux.js.org/faq/code-structure
